Move environment_check fixture to conftest.py
